The category support for "each" and so on doesn't really work
If the invoked method runs anything synchronously, and that includes a call to "each", then CpsDefaultGroovyMethods
end up getting used, which is not what we want.

One case where this happens is:

  @NonCPS
  void foo() {
     def sum = 0;
     [1,2,3,4].each { sum+= it; }  // <- we expect this to execute synchronously
     return sum;
  }

  // CPS-transformed function
  void bar() {
    foo();
  }

Another situation where this happens is when GroovyClassLoader decides to load other groovy scripts as classes,
as much of the Groovy compiler is implemented in Groovy.

This change breaks the following code:

  // CPS-transformed function
  void bar() {
     def sum = 0;
     [1,2,3,4].each { sum+= it; }
     assert sum==10;
  }

... because 'sum+=it' ends up getting CpsCallableInvocation and skips the loop. So I have to come back to this fix later.