Redone the CPS-aware 'each' method
Category kicks in automatically for every indirect call that happens within CategorySupport.use(). This includes such things like
workflow script accessing another class, which triggers GroovyClassLoader to load a script, which results in groovy compiler executing.

So we need to be very careful with which 'each' method should execute synchronously vs asynchronously.

This is not very easy to do, and after thinking about this, I think the best approximation is to have the CPS method call site
records what method it is about to invoke, so that the method invoked can test if it was directly called from a CPS method call site.

This ignores all the weird call routing that Groovy does (for example, "o.each" would work but not "o.invokeMethod('each')"),
but it is sufficient to pass the test cases pass, and so an OK implementation for now.

If we decide to revisit this in the future, perhaps another option is to speculatively execute the passed in closure synchronously,
and if we see the magic CpsCallableInvocation exception, switch to executing the rest in async function by decorating Continuation.