Making it easy to create a Continuable that prepends another one.
In an attempt to fix JENKINS-25495, I've tried to allow Continuable to
run what amounts to be an "interrupt handler" --- set aside what you are
doing, run something else, then come back to what was set aside, all in
a single Continuable.

I've ultimately fixed the problem differently, but I'm leaving the
result in here as these additions seem sensible on its own.