[JENKINS-25504]
The previous attempt failed because CpsStepContext can get duplicated because of persistence.

This is a weaker fix that only waits for the "primary" body execution. That is, this doesn't work correctly if a Step is like parallel step that executes multiple bodies at the same time.

That said, the current ParallelStep implementation never reports itself completed until all the bodies check in, so this code should work correctly with all the known Step impls.

Originally-Committed-As: 68de87c231d01605f61344cc3eccbafcbaf71470