When an Exception happens in a parallel reportthe first one.
The previous behaviour (not just related to this change) was dependent on
the ordering in the map as to which error was reported first when multiple
occured.
THis changes the behaviour such that the first error is the one that is
reported higher up.

Originally-Committed-As: 52189f1747768baed9c4d55bd5f64d1aef3cea4e