Call site tagging + enable mixing trusted/untrusted code
See doc/sandbox.md for the explanation of this change.

Previously, whether to sandbox or not was a decision made at the point
of thread creation, which made it impossible to mix trusted code &
untrusted code in single program.

This change fixes that by making this decision at the call site level.
In Jenkins, this enables plugins to package Groovy code that gets
CPS-transformed but runs outside sandbox.

Attention was paid to preserve backward compatibility with the persisted
form of the program, which is translated without any trusted/untrusted
call site tags.