[JENKINS-25263]
CpsBodyExecution.cancel() and CpsFlowExecution.interrupt() should work
even when a thread is not suspended due to StepExecution, but when it's
just paused briefly by safepoints.

This code also consolidates the handling of the logic into
CpsThread.stop().

Note that this change still doesn't address the case when a CPS thread
is executing for too long inside a non-CPS method. Say
Thread.sleep(10000), blocking on a monitor, or while(true) {} inside a
Java method that Pipeline Script invoked.