To handle the properties step, we need special casing for lists.
More specifically, we need to ensure that we use parentheses around
any step's parameters when they've got only one parameter and that
parameter is a list. Otherwise, Groovy just kinda looks at you funny
and tries to treat the list as an index fetch on the step,
which...doesn't make any sense. So!