Adds option to treat missing script file as an empty file
When automating project creation along with base Jenkins jobs
for various projects, the current behavior of the Groovy plugin requires
a development team to create an empty script file before they may be
ready to develop a proper pipeline in order to pass CI. This becomes
a barrier to a development team when the failure of the initial CI job
provides a negative code review on a change.

This commit adds an option to treat a missing script file as an empty
file thus allowing the CI job to pass. The default for this option is
"false" thus preserving the original behavior of the plugin.