[FIXED JENKINS-38268] Proper lexical scoping for closure variables.
Previously, any variables that hadn't been explicitly declared (via
`ClosureCallEnv.declareVariable`) in a `Closure` would end up getting
set in the captured `Env`. Which means we'd end up setting things
globally when we really didn't mean to. This changes that to only set
a variable in the captured `Env` if there's no `locals` variable
defined by that name *and* the captured `Env` has a non-null return
for `getVariableType(name)` - i.e., the variable name has already been
declared in the captured `Env`.