[JENKINS-43995] Per stage/step status
When creating a body StepEndNode for success, looks to the immediate
children to see if any have a FlowNodeStatusAction. If so, add a
FlowNodeStatusAction with the worst status of any of the children's
FlowNodeStatusAction to the body StepEndNode.

When creating a non-body StepEndNode, get the worst status from any of
its parents with BodyInvocationActions, and if non-null, add a
FlowNodeStatusAction to the non-body StepEndNode.

Note that we don't add a FlowNodeStatusAction for an abnormal outcome
in either case. The exception propagates to the body StepEndNode in
FailureAdapter, and onward from the body StepEndNode to its enclosing
non-body StepEndNode. getStatus() on one of these StepEndNodes will
return failure, as appropriate, while caught exceptions will not alter
the status of the enclosing StepEndNodes.