Add new public object2Groovy method for use in other contexts.
Specifically, I'm going to be consuming this in Declarative for a
Declarative-specific "directive generator" like the Snippet Generator,
but for Declarative directives like `options`, `triggers`, `when`, etc.

There'll eventually be a followup PR once that has been implemented in
Declarative to add a link to the Directive Generator or whatever we
end up calling it to the Pipeline Syntax sidepanel.jelly
(conditionally - i.e., if a new enough version of Declarative is
installed) but there's no point in doing that yet since I can't
actually implement it without this method being public. =)