Stop using junit in metastep tests
Newer versions of junit plugin have `junit` as an actual `Step`, so
anything expecting a metastep when using `junit` will not get a
metastep. So, let's stop using `junit` in tests in general. This is
relevant now in the PCT, will be relevant eventually when we bump the
junit dependency here as well.