JENKINS-51123: Fix deadlock in CpsFlowExecution
A deadlock can arise if multiple threads tries to access storage
simultaneously while one of them doing it from a synchronized
method. The cause is that timing operation also tries to get the
instance lock from inside the storage r/w lock.

Solution is to change the order of locking by having the timing
object wrapping the call to the r/w storage lock instead of the
opposite.