[JENKINS-55287] Improve error messages for nonresumable Pipelines
Also fixes tracking of whether PERFORMANCE_OPTIMIZED Pipelines are
able to resume or not. Previously, they were always marked as
resumable because right after adding the FlowStartNode, the build
was saved, cuasing the persistedClean flag to be set to true.

Now we only modify persistedClean when the build has already
completed or when Jenkins is shutting down.