[JENKINS-75067] Unexport step bodies when completed in synchronous mode
Previously, the unexport of a step's body would only happen when
scheduling the next execution (async mode).  If the step was not run
async (due to having an outcome set prior to the step starting), the
body was not removed until the CpsThreadGroup was finishing.  At this
point the body's presence was treated as a resource leak and logged a
warning.  Doing an explicit unexport for the sync case at the proper
time eliminates the warnings from the fall-back cleanup.